src/main/java/com/pragma/usuarios/web/controller/UsuarioController.java
package com.pragma.usuarios.web.controller;

import com.pragma.usuarios.application.service.UsuarioService;
import com.pragma.usuarios.domain.model.Usuario;
import com.pragma.usuarios.web.dto.UsuarioDTO;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import jakarta.validation.Valid;

import java.util.Optional;

@RestController
@RequestMapping("/")
@Validated
public class UsuarioController {

    @Autowired
    private UsuarioService usuarioService;

    @PostMapping("/propietario")
    public ResponseEntity<Usuario> crearPropietario(@Valid @RequestBody Usuario usuario) {
        Usuario nuevoPropietario = usuarioService.crearPropietario(usuario);
        return new ResponseEntity<>(nuevoPropietario, HttpStatus.CREATED);
    }

    @GetMapping("/{id}")
    public ResponseEntity<UsuarioDTO> obtenerUsuarioPorId(@PathVariable Long id) {
        Optional<Usuario> usuarioOpt = usuarioService.obtenerUsuarioPorId(id);

        if (usuarioOpt.isPresent()) {
            Usuario usuario = usuarioOpt.get();
            UsuarioDTO usuarioDTO = new UsuarioDTO(usuario.getId(), usuario.getNombre(), usuario.getRol().toString());
            return ResponseEntity.ok(usuarioDTO);
        } else {
            return ResponseEntity.notFound().build();
        }
    }
}

